Reorg for tests
