+++ .gitignore
@@ -1,6 +1,4 @@
 # Binaries for programs and plugins
-*.exe
-*.dll
 *.so
 *.dylib
 
@@ -12,3 +10,6 @@
 
 # Project-local glide cache, RE: https://github.com/Masterminds/glide/issues/736
 .glide/
+
+# VS Code
+.vscode/
